Finish off reprojection

+++ src/type.ts
@@ -1,9 +1,12 @@
 import * as arrow from "apache-arrow";
 
-export type InterleavedCoord = arrow.FixedSizeList<arrow.Float64>;
+// Note: this apparently has to be arrow.Float and not arrow.Float64 to ensure
+// that recreating a data instance with arrow.makeData type checks using the
+// input's data type.
+export type InterleavedCoord = arrow.FixedSizeList<arrow.Float>;
 export type SeparatedCoord = arrow.Struct<{
-  x: arrow.Float64;
-  y: arrow.Float64;
+  x: arrow.Float;
+  y: arrow.Float;
 }>;
 // TODO: support separated coords
 export type Coord = InterleavedCoord; // | SeparatedCoord;
